refactor : 패키기 구조 리팩터링.
