+++ src/main/java/subway/domain/station/Station.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.domain.station;
 
 public class Station {
     private Long id;
