--- src/main/java/subway/exceptions/BadSectionsForDeleteCreation.java
@@ -1,8 +1,0 @@
-package subway.exceptions;
-
-public class BadSectionsForDeleteCreation extends RuntimeException {
-
-    public BadSectionsForDeleteCreation(String message) {
-       super(message);
-    }
-}
