+++ src/main/java/subway/service/LineService.java
@@ -1,9 +1,11 @@
-package subway.line;
+package subway.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.section.SectionDto;
-import subway.section.SectionService;
+import subway.dao.LineDao;
+import subway.domain.line.Line;
+import subway.dto.LineDto;
+import subway.dto.SectionDto;
 
 import java.util.List;
