+++ src/main/java/subway/service/StationService.java
@@ -1,6 +1,8 @@
-package subway.station;
+package subway.service;
 
 import org.springframework.stereotype.Service;
+import subway.dao.StationDao;
+import subway.domain.station.Station;
 
 import java.util.List;
