+++ src/test/java/subway/line/LineDtoTest.java
@@ -3,6 +3,8 @@ package subway.line;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.dto.LineDto;
+import subway.http.request.LineRequest;
 
 @DisplayName("LineDto 클래스")
 public class LineDtoTest {
